fzj imports: various tweaks and documentation fixes

+++ lib/reda/importers/eit_version_2013.py
@@ -93,7 +93,13 @@ def _extract_emd(mat, **kwargs):
 
     # sort current injections
     condition = df['a'] > df['b']
-    df.loc[condition, ['a', 'b']] = df.loc[condition, ['b', 'a']].values
+    df.loc[condition, ['a', 'b']] = df.loc[
+        condition, ['b', 'a']
+    ].values.astype(int)
+    # for some reason we lose the integer casting of a and b here
+    df['a'] = df['a'].astype(int)
+    df['b'] = df['b'].astype(int)
+
     # change sign because we changed A and B
     df.loc[condition, ['Z1', 'Z2', 'Z3']] *= -1
